requirements.txt
# Core Framework
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-groq>=0.1.0

# Document Parsing
unstructured[pdf,docx]>=0.14.0
python-docx>=1.1.0
pdfminer.six>=20221105
pytesseract>=0.3.10
pdf2image>=1.17.0
pillow>=10.0.0

# Embeddings & Vector Store
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0

# BM25 / Sparse Retrieval
rank-bm25>=0.2.2

# Re-ranking
# cross-encoder is part of sentence-transformers

# LLM (via langchain-groq — wraps groq SDK for LLaMA 3)

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0

# Utilities
python-dotenv>=1.0.1
tqdm>=4.66.0
loguru>=0.7.2
httpx>=0.27.0
aiofiles>=23.2.1
